requirements.txt
# Enterprise Insider Threat Detection Platform
# Python 3.9+ required

# Web Framework
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9

# Database
sqlalchemy==2.0.25
aiosqlite==0.19.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# Pydantic & Settings
pydantic==2.6.1
pydantic-settings==2.1.0

# Machine Learning
scikit-learn==1.4.0
numpy==1.26.4
pandas==2.2.0

# NLP & Explainability
sentence-transformers==2.3.1
transformers==4.37.2
torch==2.2.0
shap==0.44.1
lime==0.2.0.1

# File Processing
python-magic==0.4.27
chardet==5.2.0

# Utilities
python-dateutil==2.8.2
httpx==0.26.0

# Development & Testing
pytest==8.0.0
pytest-asyncio==0.23.4
httpx==0.26.0

# Optional: For production deployment
gunicorn==21.2.0
